Enables the -t/--timeout parameter to actually.. work. Closes #4

+++ src/headers/read_input.h
@@ -13,7 +13,7 @@ typedef struct argValues {
   double* criticalMax;
   double* criticalMin;
   int* criticalInclusive;
-  int timeout;
+  long timeout;
 } argValues;
 
 struct argValues* argVals;
